Set mock api to get random pokemons
