+++ src/containers/ItemSelected/ItemSelected.js
@@ -3,7 +3,7 @@
 /* eslint-disable react/prefer-stateless-function */
 /* eslint-disable no-unused-vars */
 import React from 'react';
-import fetchPokemon from '../../services/api';
+import axios from 'axios';
 
 class ItemSelected extends React.Component {
   constructor(props) {
@@ -21,19 +21,61 @@ class ItemSelected extends React.Component {
 
   componentDidMount() {
     console.log('componentDidMount');
-    const pikachu = fetchPokemon(25);
-    console.log(typeof (pikachu));
+
+    const fetchPokemon = async (id = '') => {
+      const response = await axios.get(`https://pokeapi.co/api/v2/pokemon/${id}`);
+      const result = await response.data;
+      console.log(result);
+      this.setState({
+        id: result.id,
+        name: result.name.toUpperCase(),
+        height: result.height,
+        weight: result.weight,
+        image: result.sprites.other.home.front_default,
+        hp: result.stats[0].base_stat,
+        type: result.types[0].type.name.toUpperCase(),
+      });
+    };
+    fetchPokemon(Math.floor(Math.random() * 150));
   }
 
   render() {
     console.log('render');
     return (
-      <div>
-        <h1>
-          testing:
-          {this.state.name}
-        </h1>
+      <div className="ItemSelected col-4">
+        <div className="TopSection">
+          <figure className="TopImage">
+            <img src={this.state.image} alt="Pokemon" />
+            <figcaption>
+              Name:&#160;
+              {this.state.name}
+              <br />
+              <br />
+              Id:&#160;
+              {this.state.id}
+            </figcaption>
+          </figure>
+        </div>
+        <div className="MidSection">
+          <p>
+            Type:&#160;
+            {this.state.type}
+          </p>
+          <p>
+            Height:&#160;
+            {this.state.height}
+          </p>
+          <p>
+            Weight:&#160;
+            {this.state.weight}
+          </p>
+          <p>
+            hp:&#160;
+            {this.state.hp}
+          </p>
+        </div>
       </div>
+
     );
   }
 }
